refactor: let pdfhelper parse args only
